+++ src/declog/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version
 
-from .logging import *  # noqa F401
+from .core import *  # noqa F401
 
 
 __version__ = version("declog")
